+++ packages/dartvel_cli/test/account_page_routes_test.dart
@@ -237,7 +237,7 @@ void main() {
     File(p.join(root.path, 'pubspec.yaml')).writeAsStringSync('''
 name: refused_pages
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dartvel:
   auth:
     pages: