+++ packages/dartvel_cli/test/account_wiring_backend_test.dart
@@ -307,7 +307,7 @@ void main() {
 name: account_wiring_probe
 publish_to: none
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dependencies:
   dartvel_core:
     path: ${p.join(packages, 'dartvel_core')}