+++ packages/dartvel_cli/test/adoption_build_checks_test.dart
@@ -246,7 +246,7 @@ class _User {
 name: adopt_probe
 publish_to: none
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dartvel:
   prodBackendHost: https://example.com
 ''');