+++ packages/dartvel_cli/test/crash_ingest_backend_test.dart
@@ -145,7 +145,7 @@ Future<Directory> backendProject(String packages, String crashes) async {
 name: ingest_probe
 publish_to: none
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dependencies:
   dartvel_core:
     path: ${p.join(packages, 'dartvel_core')}