+++ packages/dartvel_cli/test/model_list_shadow_test.dart
@@ -67,7 +67,7 @@ class _Article {
 const String _pubspec = '''
 name: shadow_app
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dartvel:
   search:
     synonyms: