+++ packages/dartvel_cli/test/platform_api_oauth_backend_test.dart
@@ -313,7 +313,7 @@ void main() {
 name: oauth_probe
 publish_to: none
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dependencies:
   crypto: any
   dartvel_core: