+++ packages/dartvel_cli/test/scaffold_backend_serves_test.dart
@@ -54,7 +54,7 @@ void main() {
 name: scaffold_backend_probe
 publish_to: none
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dependencies:
   dartvel_core:
     path: ${p.join(packages, 'dartvel_core')}