+++ packages/dartvel_cli/test/server_binary_build_test.dart
@@ -79,7 +79,7 @@ void main() {
 name: server_binary_probe
 publish_to: none
 environment:
-  sdk: ^3.12.0
+  sdk: ^3.13.0
 dependencies:
   dartvel_core:
     path: ${p.join(packages, 'dartvel_core')}